chore : [#8] Id, Seq 데이터 타입 Integer로 변경(개인 앱이라는 점 고려)

+++ BE/src/main/java/dev/whatevernote/be/service/NoteService.java
@@ -20,7 +20,7 @@ public class NoteService {
 		this.noteRepository = noteRepository;
 	}
 
-	public NoteResponseDto findById(final Long noteId) {
+	public NoteResponseDto findById(final Integer noteId) {
 		return null;
 	}
